feat(crypto): derive storage encryption key
Panther wallet will be required to store data in the user browser. A major part
of this data must be encrypted. The main goal of this patch is to
deterministically derive a "storage encryption key"  that will be used to
encrypt/decrypt the data. It achieves this by modifying the existing
`deriveRootKeypairs()` to return an extra key from the user signature. The key
is derived from the module Baby JubJub of the poseidon hash of the hashed
signature `mbjj(p(p(sig)))`.

Issue-URL: https://app.clickup.com/t/866a68d9t