refactor(dapp): reorganize types folder
Without this patch, /types folder is not well
organized

This is a problem because the current Structure is not intuitive or
easy to maintain.

This patch solves the problem by improving the organization and
maintainability of the /types folder by separating types based on their
source (vendors, Panther protocol, or JS packages without types) with a
clear and intuitive structure.

Issue-URL: https://app.clickup.com/t/865c4ht1r