chore(contracts): deterministic deployment scripts
Without this patch, no utility function exists for deterministic
deployment.

This patch solves the problem by adding 3 utility functions:
`setDeterministicDeploymentProxy`,
`arbitraryDataDeterministicDeploy` and `deterministicDeploy`

The first one sets the bytecode of the deployer contract at a specific
address `0x4e59b44847b379578588920cA78FbF26c0B4956C` using
`hardhat_setCode` command. The address is the same on all of the
EVM-based networks hence we use it on the local network to keep
consistency.

The other 2 functions uses the Deployer proxy contract to deploy the
bytecode or any arbitrary data in a deterministic manner.

Issue-URL: https://app.clickup.com/t/862jwvq0j