feat(crypto): three-tiered BusTree
This commit introduces the implementation of a three-tiered Merkle tree,
named BusTree. BusTree is responsible for generating and managing
three-tiered Merkle tree that is composed of UTXO pack/queue, branch,
and root trees.

                            Root Tree
                             /  |  \
                            /   |   \
                           /    |    \
                          /     |     \
                    Branch Tree  |   Branch Tree
                       /  |  \   |     /  |  \
                      /   |   \  |    /   |   \
                     /    |    \ |   /    |    \
                UTXO Pack | UTXO Pack | UTXO Pack
                   /  |  \  |    /  |  \  |   /  |  \
                  L   L   L |   L   L   L |  L   L   L
                  e   e   e |   e   e   e |  e   e   e
                  a   a   a |   a   a   a |  a   a   a
                  f   f   f |   f   f   f |  f   f   f

This tree has three levels:
- The root tree: It is composed of branches (Branch Trees);
- Branch tree: Each branch is composed of UTXO packs (UTXO Pack Trees);
- UTXO Pack tree: Each UTXO pack is composed of leaves (transactions);

Please note that the above diagram is an oversimplified representation
for better understanding. In reality, the depth of the tree can vary
depending on the parameters provided while constructing the BusTree
object, and each tree can potentially have 2^n nodes, where n is the
depth of the tree.

Issue-URL: https://app.clickup.com/t/865ccm584