feat(dapp): zk proof generator in the browser
Without this patch, we will not be able to generate proofs in the
browser.

This is a problem because do not have implemented proof generation is
browser.

This patch solves the problem by implementing a logic to use downloaded
files and based on inputs (priv/public) to generate a proper proof,
also we extend logic with a function to do a proof validation.
This implementation is using our groth16 implementation to be compatible
with our circuits implementations.

Issue-URL: https://app.clickup.com/t/31euaqf