feat(contracts): modify simulation adding utxo
Without this patch, the reward is defined as a parameter. It means
anyone who adds UTXOs is able to put any desired value as reward.

This is a problem because at testing stage, the miners are supposed to
add UTXOs hence they can define the reward value.
since the reward value will be exchangeable for real ZKP, then the
miner may define huge rewards in order to be able to exchange these
test rewards (ZKP on mumbai) with real ZKP (ZKP on polygon).

This patch solves the problem by modifying the
`simulateAddUtxosToBusQueue` to calculate `utxos[]` and `reward`
internally without getting any parameter from external.
The reward is calculated based on a formula defined in the contract.

Issue-URL: https://app.clickup.com/t/865cpk7mw