feat(contracts:ProvidersKeys): keyrings mapped to its IDs and renaming
1. Before this commit keyrings were mapped to the addresses of operators
which is suboptimal since the operator address (but not the ID of a
keyring) may be updated.
Now a keyring is mapped to it's ID (and the operator address is a part
of the keyring record).
2. Poseidon hashers with "input in SNARK field" checks are moved to a
separate lib to share the code.
3. Indexes in the PantherForest and StaticTree made shared constants.
4. UtilsLib added to re-use shared functions. Unlike `contract Utils`,
the lib does not result in the entire contract code (incl. unused fns)
to be added to the contract which use these shared fns.
5. Furthermore, names of variables and comments clarified.
6. TODOs added.