feat(contracts:PantherForest): root history fixed
Before this commit,the logic of Forest roots caching was broken.
With this commit, `PantherForest` inherits the `CachedRoots` contract.
The latest implements fixed Forest tree past roots caching.

The new root of the Forest tree to be cached every time any of the Forest
tree's leafs gets updated. But if the Static tree is updated, all the
cache must be reset: only the latest root (generated by the Static tree
root update) should remain in the cache.