fix(contracts) magicalConstraint made fn input
Before this commit the "magical constraint" circuit public signal was
a hard-coded constant, which is incorrect: this signal must get random
value on every invocation.
This commit implements this requirement.
To avoid the "stack to deep" error, public signals are made an array.
Furthermore, more checks on public signals added.

In addition to that, error messages are moved to a separate file.