test(contracts) extend `expiryTime` of vPool
Without this patch, the expiry time is too low, hence sometimes it will
be less than current time and the contract throws because of that.
Also, the time is read by js native `Date()` so that it may differ from
the hardhat network block time.

This patch solves the problem by extending the expiry time and reading
the timestamp from the current block.