feat(contracts): add start time to bus tree
Without this patch, The START_TIME is an immutable value which needs to
be defined at deployment.

This is a problem because the contract runs behind an upgradable
proxy, and each time the contract is upgraded the constructor sets the
new timestamp, while this value must remain unchanged on upgrades.

This patch solves the problem by defining this value at `updateParams()`
only if it's undefined.
`updateParams()` sets the parameters' values at the proxy storage.