fix(contracts): remove utxo time from constructor
Without this patch, the `lastUtxoSimulationTimestamp` is defined at
deployment in the busTree storage.

This is a problem because the contract runs behind an upgradable proxy
so that the storage parameters must to be defined in the proxy
storage.

This patch solves the problem by removing the parameter from
the constructor.
This value will be updated in the proxy storage
automatically after each utxo insertion simulation.