feat(contracts): add vault address to converter
Without this patch, the prpConverter smart contract does not know the
vault address.

This is a problem because the prpConverter needs to increase the vault
allowance so that vault can transfer (aka lock)
zkp tokens from converter.

This patch solves the problem by adding the VAULT constant variable in the
prpConverter smart contract.
The vault allowance is increased each time
`prpConverter::updateZkpReserve()` is executed.
updateZkpReserve() can be executed by anyone. This function checks if
the zkp balance is more than zkp reserves, if so, it updates the
reserves, increase the vault allowance and also burn PRPs
(simulating swapping ZKP for PRP).