feat(dapp): zAccount UTXO scanning
This commit introduces a temporary delay of 5 seconds for the
initialization of the worker in the zk-proof-engine. The delay aims to
demonstrate that the worker requires a certain amount of time to
initialize properly.

This commit refactors the IPFS initialization of the Downloader in the
dapp. The previous implementation was failing to start the Downloader on
IPFS due to an incorrect worker location.

To resolve this, the worker's location is now defined during the
initialization process. In addition, the SharedWorker has been replaced
with a WebWorker. The WebWorker is a better fit for our use case as it
eliminates the need to call the downloader from different locations,
thus enhancing the application's efficiency.

Subgraph response handling in the dapp has been improved to enhance
reliability and error handling. Previously, the system would directly
return the branchBatches response, which could lead to potential issues
due to unexpected response formats or errors. The new approach ensures
the response is correctly formatted as a GetBranchBatchesResponse before
returning it.

Additionally, debug information has been added to the zAccount proof
calculation process in development mode. This change is intended to aid
in troubleshooting and provide a clearer understanding of the process.

This commit focuses on enhancing the readability and maintainability of
the codebase. Previously, the codebase contained large and complex
functions that were challenging to debug, test, and understand, creating
a bottleneck in the development process and increasing the potential for
errors.

To address this, the commit breaks down these large functions into
smaller, more manageable ones and introduces helper functions to
eliminate redundancy and improve code clarity. This refactoring makes
the code easier to read, understand, and maintain, thereby accelerating
development and reducing the likelihood of errors.

Key changes include:
-  Large functions have been refactored into smaller ones for improved
   readability and maintainability.
-  `GetBranchBatchesResponse` has been renamed to `GetBatchesResponse`
   for better clarity.
-  Comments indicating future work needed for further refactoring and
   cleanup have been added.
-  Complex logic has been extracted into helper functions
   `decodeAndFilterNotes`, `constructZAccountUtxo`, and
   `constructDecodedNoteDictionary` in `sync.ts`.