chore(contracts): fix bus tree interface
Without this patch, the name of parameters which are returned by
`addUtxosToBusQueue` and `addUtxoToBusQueue` are wrong.

Since the `addUtxosToBusQueue` adds more than one utxos in
the queue, so that it returns `firstUtxoQueueId` and
`firstUtxoIndexInQueue`.
In the other hand the `addUtxoToBusQueue` adds one utxo in the queue, so
that it returns `queueId` and `indexInQueue`.