refactor(root): use rimraf to clean workspaces
Without this patch, we will use the rm -rf command
to remove all "node_modules" from all workspaces.

This is a problem because we will have to manually extend
the command to include new workspaces and it doesn't work
with Windows.

This patch solves the problem by using the `rimraf` package
and run it using yarn