feat(contracts): validate circuit pointer
Without this patch, the contract does not validate if the circuits for
accounting prp and conversion of prp have been defined.
Also, Id does not check the address that execute `accountPrp()`.

This patch solves the problem by adding the validations in the
`accountPrp()` and the `accountPrpConvertion` to check circuit ids.
`accountPrp()` has extra validation to check the caller
(aka`msg.sender`).
The `accountPrpConvertion` can be executed by anyone who approve the
`Vault`.