feat(contracts): StealthDelegate implemented
It executes multicalls from deterministically computed address of the
"stealth" contract.

StealthDelegator is an unusual contract: just like any other contract,
it has the creation code, but its runtime bytecode is never deployed.
It is "delegator" since it executes the DELAGATECALL with given params.
It is "stealth" as it runs (when called via CREATE or CREATE2) during
the "deployment" only, and then it "disappears".

When the "deployer" (the account executing CREATE or CREATE2) calls this
creation code, the code does NOT create any new contracts, but instead:
- executes the DELEGATECALL with the given `to` and `data`, and then ...
- self-destructs, sending remaining native token balance to the deployer.

Users may be sure that, inside the DELEGATECALL:
- caller (`msg.sender`) is always the deployer
- executing account (`address(this)`), if the deployer executes CREATE2,
  may be deterministically computed before the "deployment".

We are going to use this contract as a "universal plugin" for so-called
DeFi integration.