feat(subgraph): Introduce forest root tracking
The application previously lacked the ability to monitor updates to the
forest root in the subgraph, which impeded our capacity to process
onchain transactions with historical roots.

To address this, a new feature has been implemented to track the forest
root in the subgraph. This involves the addition of a new event to
PantherPoolV1.json and the creation of a new entity in schema.graphql. A
new function to handle the RootUpdated event has been added in
handePoolContract.ts and updates to subgraph.yaml have been made to
accommodate this event. The dApp env file has also been updated with the
new deployed subgraph QID.