feat(dapp): fixed issue with overlapping modals
Without this patch, we will have a problem with rendering modals over
each other, "risk and disclaimer" modal should be prioritized.

This is a problem because we didn't get in consideration that has user
accepted "risk and disclaimer" and we show it and we show connecting to
the wallet modal

This patch solves the problem by hiding the modal when the user has
not accepted "risk and disclaimer" and right after accepting we show
connection modal.

Issue-URL: areTermsAccepted