fix(contracts): accept private message on claim
Without this patch, the `PrpVoucherGrantor::claimRewards` does not
accept the private message.

This is a problem because private message should be part of tx note
event and it needs to be received from dapp.

This patch solves the problem by changing the `claimRewards()` to accept
the privateMessages param.

The privateMessages contains the zAccount utxo data, hence it must start
with `0x06` which is the message type for `UTXO_ZACCOUNT`.