fix(contracts): fix prp conversion tx note
Without this patch, the prp conversion does not accept the private
message. The order of tx note params is also wrong. the zAccount utxo
commitment should be inside the `MT_UTXO_BUSTREE_IDS` type.

This is a problem because the private message must be computed in dapp.
so the function needs to accept it as a parameter.
The private message contains two message type. `MT_UTXO_ZACCOUNT` and
`MT_UTXO_ZASSET_PRIV`.