feat(crypto): Add support for PRP exchange transaction note type 3
Without this patch, our crypto module lacked the capability to handle
PRP exchange transaction note type 3, which was limiting our ability to
fully interact with the PRP exchange.

This patch solves the problem by adding support for transaction note
type 3 in the crypto module. The changes include the addition of new
constants, updating the `decodeZAccountActivationTxNote` function, and
adding a new `decodeTxNoteType3` function. Additionally, the patch also
includes the creation of a new `TxNoteType3` type.

Issue-URL: https://app.clickup.com/t/8692t8m60