fix(contracts): fix cached root index
Without this patch, the `getCacheStats` returns an error if the root
history is full.

This is a problem because this function is supposed to return
`numRootsCached` and `latestCacheIndex`.
Since the history size is 256, so if the history
is full, the `latestCacheIndex` should be 255, but the contract reverts.

This patch solves the problem by returning 255 as `latestCacheIndex` if
the history is full.

Issue-URL: https://app.clickup.com/t/86930wezp