fix(circuits): keys and wasm files in amm and zAccountRegistration
This update ensures that the keys and wasm files in the amm and
zAccountRegistration modules are kept up-to-date. Previously, the lack
of this update could lead to outdated files which may affect the
performance and reliability of the system. By applying this patch, we
mitigate the potential issues associated with outdated files.

Issue-URL: https://app.clickup.com/t/8692t46k6