feat(contracts):  generate rewards by msg relayer
Without this patch, there is no logic to let msg relayer generate
prp rewards.

This is a problem because the relayer is the contract that keep track
of zkp bridge events. So it needs to be able to generate prp rewards for
the user who release ZKPs from vPool to prpConverter.