feat(crypto): Add Commitment Message Encryption for UTXO's hashes
Introduces Commitment Message Encryption to handle spent or received
UTXO's commitment hashes, which are critical for rebuilding past
transactions. The previous approach required the root spending private
key to calculate the commitment hash, presenting a security
vulnerability. This change mitigates that risk by using commitment
messages.

The motivation for this change is to enhance security by eliminating the
need for access to the root spending private key when calculating
commitment hashes.

Issue-URL: https://app.clickup.com/t/86935up5e