feat(dapp): Integrate zAsset Commitment Message into PRP Conversion Note
This commit integrates zAsset Commitment Message into the transaction
note of PRP Conversion. This is crucial for reconstructing past
transactions. Previously, the root spending private key was required for
the calculation of the commitment hash, which posed a security
vulnerability. This update addresses this risk by utilizing commitment
messages instead.

The primary motivation for this change is to enhance security. By
incorporating commitment messages, we eliminate the need for access to
the root spending private key for calculating commitment hashes.

This change improves the overall security of transaction processing and
reduces potential risks associated with private key access.

Issue-URL: https://app.clickup.com/t/86935up5e