refactor(crypto): streamline message handling
Replace UTXOMessage with a more generic Message type. This change is
intended to simplify the codebase and improve maintainability. The new
Message type is more flexible and can handle more types of messages,
reducing the need for multiple similar types. This will make the code
easier to work with in the future.

The UTXOMessageType enum has been replaced with the MessageType enum,
which is used throughout the code to identify the type of a Message.
This change has been reflected in all parts of the code that previously
used UTXOMessageType.

The UTXOMessage type has been removed entirely, as it is no longer
needed. All references to UTXOMessage in the code have been replaced
with Message.

The structure of the Message type has been changed to be more flexible
and easier to work with. It now allows for optional keys, which makes it
easier to work with messages that may not have all possible fields.

The ZAccountUTXOMessage, ZAssetUTXOMessage, ZAssetPrivUTXOMessage, and
CommitmentMessage types have been combined into a single PrivateMessage
type. This simplifies the code and makes it easier to handle private
messages.

All of these changes are aimed at improving the quality and
maintainability of the code. They should not have any impact on the
functionality of the application.