feat(contracts): MT_SPENT_2UXTO msg type added
This new message type (emitted with a MASP transaction) helps the DApp
to reconstruct the history of MASP transactions by providing commitments
to the UTXOs spent in a transaction (encrypted with the encryption key
of the spender). Without this message, detection of spent UTXOs (of
user) requires scanning of blockchain events "since the beginning up
until now" as the DApp needs to know all UTXOs of a user, in order to
detect which ones of those UTXOs are spent. With the new message added
the DApp does not need to know all UTXOs of a user in order to detect
spending UTXOs, so the efficient scanning "from now back to the past"
may be used.