fix(circuits): rule and leafid computation
Without this patch inclusion prover will not work as the addition of
both kyc and kyt rule id is resulting in a number that can not be
contained by a 24 bit variable.

With this patch the 24 bit variable is increased to 240 (24 * 10).
Since we can have 10 leaves of 24 bits (ruleid+leafid) the parameter
which is the bit size should also be updated in order to capture the
growing number.

Issue-URL: https://app.clickup.com/t/86939bb8d