chore(dapp): update CODEOWNERS file
This commit refines the CODEOWNERS file to better reflect the current
responsibilities of the team. The motivation behind this change is to
ensure that the right individuals are notified and involved in the
review process for certain areas of the codebase.

The following changes have been made:

-  Added more specific paths to @oleksii9450320, including
   'dapp/webpack.config.js, 'dapp/tsconfig.json' and
   'dapp/babel.config.json' config
-  Expanded @neuodev's areas to include 'dapp/src/redux/*' and
   'dapp/src/hooks/*'
-  Extended @dcheglakov's areas to encompass '.storybook/*', 'public/*',
   and specific frontend files and styles

By refining the CODEOWNERS file, we aim to streamline our code review
process and ensure that changes are reviewed by the most relevant team
members.