chore(root): Update CODEOWNERS file
This commit updates the CODEOWNERS file to reflect changes in the
project structure and responsibilities.

The changes include:
-  Simplified directory references for @pycckuu, @oleksii9450320,
   @neuodev, and @dcheglakov
-  Introduced *.scss and *.tsx file ownership for @dcheglakov

The update is crucial to ensure that the right individuals are notified
for changes in their areas of responsibility, improving the efficiency
of code reviews and maintaining the quality of the codebase.