fix(subgraph): resolve unique constraint violation on RootUpdated entity
This commit resolves a non-deterministic error caused by violation of
unique constraint on the root_updated_id_key. The error occurred due to
the use of updatedRoot as the id for the RootUpdated entity. This is not
unique because multiple RootUpdated events can have the same
updatedRoot, which should not happen in production.

To fix this, the transaction hash and log index are now used as the id
for the RootUpdated entity. This ensures uniqueness and prevents the
error from recurring.

Issue-URL: https://app.clickup.com/t/8693ad3z0