refactor(contracts): StealthExec refactored
It is made a `library` and renamed since the `library`, rather than the
`abstract contract`, better organizes and explains the code.
Then, SELFDESTROY opcode removed to prepare for the EIP-4758 (EIP-6780
and similar EIPs).

StealthExec allows to `CALL` from a "stealth" account defined by the "salt".
To execute the CALL with the caller address (`msg.sender`) different than
the account invoking this library, the library code "deploys" at that address
a special contract, "StealthCaller", the creation code ("init code") of which
executes the CALL.
The StealthCaller is an unusual ("stealth") contract. Its "deployed code" does
not exist - the creation code leaves no bytecode at all to deploy on the chain.
The `CALL` execution, rather than the runtime code deployment, is the purpose
of the creation code invocation.
The caller (`msg.sender`) inside the `CALL` is the address of the StealthCaller.
This address is computable upfront and linked to the address to be CALLed and
the call data - thanks to the `CREATE2` that invokes the creation code.