feat(contracts): StealthEthPull implemented
It let "pull" ETH from a deterministically defined "stealth" account via
a special creation code (init code). The library is added as it is will
be used for depositing the native token (ETH, Matic, etc..) to the MASP.

The library allows "oulling" the entire ETH balance from a deterministic
"stealth" account. In order to "pull" ETH from that account, it deploys
at that account address the "SelfDestructor" contract.
The later self-destructs, causing transfer of ETH (remained before the
deployment) from that address to the "deployer".
The SelfDestructor is not a "normal" contract - it has no runtime code at
all. Its creation code (init code) does nothing but self-destroys the own
context. The context destruction causes the transfer, which is the purpose
of the SelfDestructor "deployment".
Thanks to CREATE2 invoking the creation code, the SelfDestructor address
may be computed in advance.

The code is adopted to the EIP-6780 (it self-destruct in the same tx with
the "deployment").