feat(contracts): MultiDelegatecall implemented
If used together with the Multicall, this contract allows atomic execution
of a few contract calls (so that, either all calls will be executed together
or no calls will be executed) with the `msg.sender` inside these calls, being
the account that calls the `function delegatecall` on this contract (it may
be called via a proxy only).

This feature is required to safely approve a token for some contract and, in
the same blockchain transaction, trigger that contract to pull the token w/o
the risk of front-running.