feat(contracts): extend private msg validation
This patch adds a private helper function in
the `TransactionNoteEmitter` contract.
This function checks if:
1. The private msg has a valid length.
2. The private msg has the valid data.

For example, the private msg for zAccountActivation must
start with `0x06` followed by 96 bytes. So, the newly
added helper
function `_sanitizePrivateMessage` checks the length and
the first bytes to be started with 0x06 (which is
the msg type for utxo zAccount).